[WLD-150] build tag for unit teests added
