+++ oracle-app/internal/exchange/binance/helpers_test.go
@@ -1,3 +1,5 @@
+// +build unit
+
 package binance
 
 import (
