+++ oracle-app/internal/utils/utils_test.go
@@ -1,3 +1,5 @@
+// +build unit
+
 package utils
 
 import (
