+++ x/currencies/msgs/issue_currency_test.go
@@ -1,3 +1,5 @@
+// +build unit
+
 package msgs
 
 import (
