+++ x/currencies/queries/currency_res_test.go
@@ -1,3 +1,5 @@
+// +build unit
+
 package queries
 
 import (
