+++ x/currencies/queries/destroy_res_test.go
@@ -1,3 +1,5 @@
+// +build unit
+
 package queries
 
 import (
