+++ x/poa/msgs/common_test.go
@@ -1,3 +1,5 @@
+// +build unit
+
 package msgs
 
 import (
