+++ x/poa/msgs/remove_validator_test.go
@@ -1,3 +1,5 @@
+// +build unit
+
 package msgs
 
 import (
