+++ x/vm/internal/keeper/common_test.go
@@ -1,3 +1,5 @@
+// +build unit integ
+
 package keeper
 
 import (
