+++ x/vm/internal/keeper/keeper_test.go
@@ -1,3 +1,5 @@
+// +build unit
+
 package keeper
 
 import (
