+++ x/vm/internal/keeper/server_test.go
@@ -1,3 +1,5 @@
+// +build unit
+
 package keeper
 
 import (
