+++ x/vm/internal/keeper/vm_genesis_integ_test.go
@@ -1,3 +1,5 @@
+// +build integ
+
 package keeper
 
 import (
