+++ x/vm/internal/keeper/vm_grpc_test.go
@@ -1,3 +1,5 @@
+// +build unit
+
 package keeper
 
 import (
