+++ x/vm/internal/types/events_test.go
@@ -1,3 +1,5 @@
+// +build unit
+
 package types
 
 import (
